Rearrange contact files
