+++ src/CabanaPD.hpp
@@ -16,7 +16,6 @@
 #include <CabanaPD_Boundary.hpp>
 #include <CabanaPD_Comm.hpp>
 #include <CabanaPD_Constants.hpp>
-#include <CabanaPD_Contact.hpp>
 #include <CabanaPD_DisplacementProfile.hpp>
 #include <CabanaPD_Fields.hpp>
 #include <CabanaPD_Force.hpp>
@@ -30,8 +29,10 @@
 #include <CabanaPD_Types.hpp>
 #include <CabanaPD_config.hpp>
 
+#include <force/CabanaPD_ForceModels_Contact.hpp>
 #include <force/CabanaPD_ForceModels_LPS.hpp>
 #include <force/CabanaPD_ForceModels_PMB.hpp>
+#include <force/CabanaPD_Force_Contact.hpp>
 #include <force/CabanaPD_Force_LPS.hpp>
 #include <force/CabanaPD_Force_PMB.hpp>
